refactor([hoist-certmeta-to-top-level-lib]): move readCertMeta to top-level lib/ + take metaPath arg
The previous server/lib/certMeta.js baked PATHS.data at module load, so
scripts/print-access-url.js could not reuse it without crossing the
scripts/server boundary and instead carried its own inline JSON.parse +
try/catch. Move the module next to lib/certPaths.js (zero-dep,
side-effect-free, the same constraint set as the existing sibling) and
change the signature to readCertMeta(metaPath) so each caller passes the
path it already resolves via certPaths(PATHS.data).meta.

Four callers updated: server/lib/peerSelfHost.js,
server/lib/networkExposure.js, server/services/certRenewer.js (two call
sites), and scripts/print-access-url.js (which drops the inline
try/catch in favor of readCertMeta(META_PATH)?.mode || ''). The test
file moves alongside the module and gains one new case asserting
metaPath is honored.

JSON.parse + try/catch is inlined rather than reusing
server/lib/fileUtils.js#safeJSONParse — the JSDoc notes this is the
cost of the zero-dep boundary that lets scripts/ import the helper.
